refactor: standardPronunciation 기능을 개별 폴더로 분리합니다
